Update patch set 1
Patch Set 1: Presubmit-Verified+1

=== Finished presubmit run: L39600000128117841 ===

Status: https://android-build.googleplex.com/builds/where-is-my-cl/android-review/569124?revision=1




TreeHugger finished with: 15 passed.







WARNING: The tip-of-tree(4503831) for git_master/gce_x86_phone-userdebug_fastbuild3c_linux may be broken.
The presubmit build passes with the last-known-good-build(4503524), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.









WARNING: The tip-of-tree(4503831) for git_master/marlin-userdebug_fastbuild_linux may be broken.
The presubmit build passes with the last-known-good-build(4503524), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.




WARNING: The tip-of-tree(4503831) for git_master/sailfish-userdebug_fastbuild3_linux may be broken.
The presubmit build passes with the last-known-good-build(4503640), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.



WARNING: The tip-of-tree(4503820) for git_mirror-aosp-master-with-vendor/aosp_sailfish-userdebug may be broken.
The presubmit build passes with the last-known-good-build(4503521), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.




I am TreeHugger, and I have hugged this change!

Have a question? Found a bug? http://go/treehugger-troubleshoot

Patch-set: 1
Tag: autogenerated:TreeHugger:Verify-Pass
Reviewer: Treehugger Robot <1062513@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Verified=+1