Use an empty string as unique id for C structs.
Some structs could be included without extern "C" guards. If the same
struct is included in a C++ source file and a C source file, it's unique
type name will be mangled in  one and its name in the other. To work
around this inconsistency, use an empty string for C structs. While
comparing two structures, if either of the old / new one has an empty
unique type id, compare their names instead.

Bug: 69568963

Test: tests/test.py

Change-Id: Ie285c4691dbaa58389a041fee5116ccbfc7985ce