gsi_util: add build.py to build and pack gsi_util
There are some command in build.py:

'setup_env' command can build the dependency files and copy these
files to the script.

'build' command also zip the dependency files to a zip file which
could be released independently.

The patch includes a list REQUIRED_ITEMS, which maintains the
source and destination of the dependency files. INCLUDE_FILES
maintains the files to be included to zip file additionally.

Uses 'list_deps' to list all dependency files.

Bug: 70253683
Test: setup_env and build command and check the files
Change-Id: I0b36da4457fdd04f9995cf30aadb26a2e32c1dd8