Monkey: Avoid ActivityManager waiting for Monkey.this
When ANR happens, ActivityManager calls ActivityController's
appNotResponding method, which waits to acquire "Monkey.this"
lock. However, this lock is also acquired by runMonkeyCycles()
method while capturing bugreports(if enabled).
This would mean that ActivityManager would be unnecessarily
blocked waiting for bugreport collection to end.
This could trigger watchdog as well since handler of
ActivityManager is blocked.
To get around this in ActivityController's appNotResponding(),
spawn a new thread that waits for "Monkey.this" and return
immediately.

Test: Build passes

Change-Id: Ie220ebe165230ea295eecfe963e856f72cdec9dd