vndk-def: Add VNDK-Private and VNDK-SP-Private
This commit adds VNDK-Private and VNDK-SP-Private.

VNDK-Private stands for the VNDK libs that can be used by other VNDK
libs but must be inaccessible to vendor modules.  It is the replacement
for VNDK-Indirect, which was defined as an alias of VNDK-Indirect
previously.

VNDK-SP-Private stands for the VNDK libs that can be used by other
VNDK-SP libs but must be inaccessible to vendor modules.  This is the
replacement for VNDK-SP-Indirect-Private.

Test: ./tests/run.py
Change-Id: I499d39e393cf265d59ec529e69ea15b670384e99