gsi_util: change verbose logs to be outputted in debug mode
Each Mounter output an error log if it cannot prepare a file,
but it could be not really an error sometimes. For example,
some files in sepolicy_checker are optional, and could got
the log:

  Fail to prepare file: vendor/etc/selinux/vendor_sepolicy.cil

The patch changes these logs to be 'INFO' level.

This patch also changes run_command() to output logs with 'ERROR'
level when log_stdout or log_stderr is True.

Also, the patch change it only outputs stack dump in debug mode.

Bug: 70588453
Test: './gsi_util.py dump --system system.img'
   and the system.img doesn't exist.
Test: './gsi_util.py check_compat --vendor adb --system adb'
   and vendor/etc/selinux/vendor_sepolicy.cil isn't in the device.

Change-Id: Ibdcb6df459f88ace9159f1a979f280452454a84f