Update patch set 5
Patch Set 5:

(1 comment)

At high-level, this CL does not look correct to me.  The correctness depends on the linkage of the inline function.  If (i) the inline function has internal linkage, (ii) the inline function has hidden visibility, or (iii) the lib is built with -Bsymbolic, it will be fine.  However, if the inline function has external linkage, the dynamic linker may link the usages in this lib to the definitions from other libs.

Patch-set: 5
Reviewer: Logan Chien <1014807@85c56323-6fa9-3386-8a01-6480fb634889>