Add --no-hidden-api-checks to runtest
New test tools are adding --no-hidden-api-checks when tests have
hidden-api-checks=false in their AndroidTest.xml, however this is not
supported by runtest. As only runtest is available in AOSP, this leaves
no option to make tests of internal classes pass.

Disable the check in runtest so we can still run tests in AOSP. This
might cause tests to pass even though they do not have configured
hidden-api-checks=false, but runtest is deprecated and apparently not a
source of truth for test success anyway.

Test: runtest frameworks-net now passes
Change-Id: I9be9674d0bf2588ff509121273379f08d4cb3aea