Update version of D8 in buildtools to latest stable version (1.0.35)
The version used to build the platform at a given point in time
might not be the best to use for external developers. Sometimes
bugs in old VMs have to be worked around in D8. Those bugs do
not affect the platform build, but could affect other users.
Therefore, we use separate versions for building the platform
and for shipping in buildtools.

Additionally, make sure that the licenses for libraries shipped
with D8 are reflected in the buildtools NOTICE file.

Test: make -j PRODUCT-sdk_phone_armv7-win_sdk sdk_repo
Test: make -j PRODUCT-sdk_arm64-sdk sdk_repo
Bug: 111748733

Change-Id: I0ca07e45647ea95816f5e96e991673ee24ed9983