Update patch set 8
Patch Set 8:

You probably need to rebase this after Andreas' change.

Given that this works around a bug in addr2line, I'm inclined to let this go through. It would be nice to move to the llvm version of this utility though.

Patch-set: 8
Reviewer: Gerrit User 1019050 <1019050@85c56323-6fa9-3386-8a01-6480fb634889>