Let header-abi-dumper output undeclared types
When the compiler finds an unknown type, it queries ExternalSemaSource
for the declaration. If the type is not declared anywhere, the compiler
ignores the type name and sets the type to int.

To keep the type name in AST, this commit adds FakeDeclSource that
creates declarations for the unknown types.

Test:
Bug: 117582158
Change-Id: I2d49b7a4f9e2a991224e6702938bc63744bba039