runtest: delete .lock in out
when run cmd fail, the alt_cmd begins to run to
rebuild all dependencies. But it fails because
the previous build job which is running cmd hasn't
exit.

The .lock should be deleted in folder out in order to ensure
the folder out will not be locked by previous build job.

BUG:
Change-Id: I8365997a363b2817b49fbff1a8154664fd2dd033
Signed-off-by: mtk13799 <xin.huang@mediatek.com>