Remove private-stub-annotations-jar from android_stubs_current et al.
This change removes private-stub-annotations-jar as a static_lib from
android_stubs_current, android_system_stubs_current, and
android_test_stubs_current. When this was here, these classes:
 - android.annotation.Nullable
 - android.annotation.NonNull
 - androidx.annotation.RecentlyNullable
 - androidx.annotation.RecentlyNonNull
but no other classes from those packages are part of the system module
when building against these stubs with -source 1.9. This means that
javac doesn't look on the classpath for other classes from those
packages, which causes build failures. This only works now because
everything that builds against them uses -source 1.8, which doesn't
have this issue (the presence of some classes from a package on the
bootclasspath doesn't stop javac looking for more classes on the
classpath). Some testing below is done with
https://r.android.com/1145829 patched in, so that they are built with
-source 1.9.

This has the effect of removing the above-listed annotations from
$ANDROID_HOST_OUT/sdk/sdk/android-sdk_eng.peteg_linux-x86/platforms/android-R/android.jar.

It doesn't affect frameworks/base/api/current.txt or any other .txt
file, as those classes are already not mentioned in any of those.

Test: m libchips (with the above patch; fails without this change)
Test: m docs (with the above patch)
Test: m sdk (without the above patch, with lunch sdk-eng)
Test: m droid docs java (without the above patch)
Bug: 142896162
Change-Id: I60bbec6644093c46cb84cbd982bbdb372cd6bf43