Update patch set 2
Patch Set 2:

(3 comments)

Thanks for uploading the WIPs!  I didn't do a detailed review yet (and I'm not sure anyway if it's worth doing a line-by-line review before you make the changes we agreed on in the design doc), but here are a few broad comments to start with.

Patch-set: 2
Reviewer: Gerrit User 1006073 <1006073@85c56323-6fa9-3386-8a01-6480fb634889>