header-checker: Downgrade to C++ 14
This commit downgrades the C++ standard from 17 to 14 because
"llvm/Support/RWMutex.h" from clang-r370808 will use
`std::shared_mutex`.  But the prebuilts are built with a different
implementation.  Furthermore, the Mac OS X 10.8 (or 10.9) does not
support `std::shared_mutex`, which results in availability compilation
errors.

This commit switches the option to `-std=c++14` to avoid potential
compilation error and fix the Mac build.

Bug: 139945549
Test: OUT_DIR=out && \
      development/vndk/tools/header-checker/android/build-prebuilts.sh

Change-Id: Ic23fbdb67f7cd0700a3001c644485f7d8d0ee3e2