Move the app to /system_ext
As mentioned in go/partitions-in-r, we plan to forbid hidden API in
product partition.
So this app should be moved to /system_ext because it uses hidden API.
It is not permanent, so they can be moved back to /product due to
needs if there isn't hidden usage anymore.

Bug: 137908189
Test: m
Test: check whether device boots successfully
Change-Id: Id1e897eadea814b780f199fa7f3d683924eb4ab3
Merged-In: Id1e897eadea814b780f199fa7f3d683924eb4ab3