Fix makefile parsing for intellij-gen.sh script
Do not assume that all modules have an Android.mk makefile
as the makefile generated in out/soong is named differently.
Use the detected makefile name if possible instead of assuming
Android.mk.

In makefile parser store values in a Set instead of appending values in
a string. This avoids duplicate values and avoid crashes due to that
String can only have 2GiB of data.

Bug: 150169583
Test: manual, execute:
    development/tools/idegen/intellij-gen.sh <project-path>
    and it does not crash due to out of memory exception.
    Generated intellij project can be loaded in android studio.
Change-Id: Id45bd891a11f35c926e248ba38b3dbd90885a980