Add product_available to VNDK snapshots
Using vndkproduct.libraries.txt, add 'product_available' property to
the selected VNDK libraries. If the snapshot does not have
vndkproduct.libraries.txt, create an empty list file and make all
snapshot libraries available to product.

For VNDK v28, auto generate the prebuilt_library modules of
libprotobuf-cpp-* that are required for the vendor compatibility to
build the vendor modules that are still using the old version of
libprotobuf-cpp libraries.

This CL also removes the code for the old VNDK snapshot v27.

Bug: 174323911
Bug: 142220143
Bug: 150902910
Bug: 159433338
Test: run update.py for various snapshot versions
Change-Id: I3f680abcbf73e3f74c5dad49f57e02de4d87a148