Fixes for cargo2android.py
Fix two issues that came up when upgrading crates.

First, the code that preserves license blocks currently also preserves
genrules if they come before a rust_ rule, causing them to be
duplicated.  Let's fix that.

Second, as an optimization, do not generate a TEST_MAPPING file (which
is slow due to Bazel queries) if we were not invoked with --tests.

Test: Upgrade some crates
Change-Id: Ie4f62cd04c2f1dd0de788db0ca5e77502fb3b4c8