Find local binaries by name
Prior to this change, attempting to find a binary failed because
the script was looking for `/proc/<pid>/exe`, which never will be
found locally.

Now we cat `/proc/<pid>/cmdline` to find the binary, then search for
that path. If found, we use that which allows locating symbols for
a binary by name.

Test: ??? TODO
Change-Id: I5b233225ca5931c6b1f2db87a7f31a4ee358d565