Split the sdk-repo platform-tools out of the main sdk build
Right now, you have to build the full legacy SDK package in order to
extract the different sdk-repo packages. As we're deprecating support
for building the system image on Mac, we'll no longer be able to build
the full system image there. It's also a bit wasteful for the Windows
SDK, which only needs to extract a small portion of the SDK.

So instead of trying to make the legacy SDK package smaller, move the
platform-tools package build into it's own make logic, not using atree.
A future change will split build-tools as well.

Bug: 187222815
Test: check that the linux/mac/windows sdk-repo-...-platform-tools-...zip
  files contain the same files
Change-Id: I7773b321fddfc478472837d5f91f02b3b16f0bb5