Extend header-abi-dumper/linker -root-dir
This commit extends the format of --root-dir to <path>:<replacement>. If
a source path starts with <path>, the prefix is replaced. The build
system can use this option to have the ABI tools replace non-default
OUT_DIR with "out". This commit also allows more than one --root-dir.
When the tool normalizes a source path, the tool finds the longest
matching root dir.

Test: ./header-checker-unittests
Bug: 192157880
Change-Id: I3a92c12a2310df5677f5691c65a388e927b80e8d